Patch Insert Table Data

+++ include/parser/tree/datatype/DataType.h
@@ -26,10 +26,9 @@ class DataType: public TreeItem {
 
 public:
     DataTypeEnum type;
-    unsigned long long size;
+    int size;
 
     static DataType *convertToDataType(std::vector<Symbol *> symbols);
-
 };
